Moved generate report buttton to sidebar. Only for admin sidebar
